requirements.txt
python-telegram-bot==21.6
playwright==1.46.0
python-dotenv==1.0.1
